Integrate selection highlighting with outline system
Replaces the old selection and target system with a new SelectionHighlighter component that uses an integrated Outline system for visual feedback. Removes the Selectable and TargetSelectSystem scripts and related assets, migrates and renames the QuickOutline implementation under InputController/Outline, and updates InputController to enable or disable highlighting on selection. Scene and prefab references are updated accordingly.
