+++ Red Strike/Assets/InputController/Outline.meta
