+++ Red Strike/Assets/InputController/Outline/Resources.meta
