+++ Red Strike/Assets/InputController/Outline/Resources/Materials.meta
