+++ Red Strike/Assets/InputController/Outline/Resources/Materials/OutlineFill.mat.meta
