+++ Red Strike/Assets/InputController/Outline/Resources/Materials/OutlineMask.mat
