+++ Red Strike/Assets/InputController/Outline/Resources/Shaders.meta
