+++ Red Strike/Assets/InputController/Outline/Resources/Shaders/OutlineFill.shader.meta
