+++ Red Strike/Assets/InputController/Outline/Resources/Shaders/OutlineMask.shader
