+++ Red Strike/Assets/InputController/Outline/Scripts.meta
