+++ Red Strike/Assets/InputController/SelectionHighlighter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d3e01160669f82e499d3f5c9fdb4eebc
