--- Red Strike/Assets/InputController/TargetSelectSystem.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-namespace InputController
-{
-    public class TargetSelectSystem : MonoBehaviour
-    {
-    }
-}
