--- Red Strike/Assets/InputController/TargetSelectSystem.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: feff55c0853c95440ae4da2a8b69b388
