+++ Red Strike/Assets/Scenes/SampleScene.unity
@@ -586,7 +586,7 @@ GameObject:
   - component: {fileID: 460729959}
   m_Layer: 6
   m_Name: Sphere
-  m_TagString: Untagged
+  m_TagString: Enemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
