--- Red Strike/Assets/Selectable.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6b8a3120469178c448f9c60d0a2bd7f5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
