--- Red Strike/Assets/Selectable/DashedLine.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2074f9565c56811498d176e7fb38eb3c
