--- Red Strike/Assets/Selectable/DashedLineMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 20debfb7cf15bcb48b9fb74ecc997944
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
