--- Red Strike/Assets/Selectable/LineArrow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d89f4cdf40f986347847aaaf8e5ce8fd
