--- Red Strike/Assets/Selectable/LineArrow.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 265d5d395b5cebf4eaafe252285f9994
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
