--- Red Strike/Assets/Selectable/SelectableEnemy.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4e6a96cc257ece14dbaa4b909156bf14
