--- Red Strike/Assets/Selectable/SelectableObject.cs
@@ -1,16 +1,0 @@
-/*
-using UnityEngine;
-
-public abstract class SelectableObject : MonoBehaviour
-{
-    public virtual void OnSelect()
-    {
-        Debug.Log(gameObject.name + " selected!");
-    }
-
-    public virtual void OnDeselect()
-    {
-        Debug.Log(gameObject.name + " deselected!");
-    }
-}
-*/
