--- Red Strike/Assets/Selectable/SelectableObject.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7e592eae79ce75842bb22c08a0b63ed0
