--- Red Strike/Assets/Selectable/SelectableVehicle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a6f8716286822ed4db451c22933417a9
