+++ Red Strike/Assets/VehicleSystem/Vehicles/OrnithopterB/OrnithopterB.prefab
@@ -361,6 +361,8 @@ GameObject:
   - component: {fileID: 4964142129493549323}
   - component: {fileID: 9172212943852637613}
   - component: {fileID: 6800533291024861918}
+  - component: {fileID: -1355217651115294554}
+  - component: {fileID: 4281797126201779825}
   m_Layer: 6
   m_Name: OrnithopterB
   m_TagString: Vehicle
@@ -461,6 +463,8 @@ MonoBehaviour:
   loiterRadius: 40
   maxBankAngle: 100
   bankSpeed: 30
+  parameterVariance: 0.1
+  altitudeVariance: 5
   barrelPoint_A: {fileID: 5423563311206911032}
   barrelPoint_B: {fileID: 3703916012389026853}
   barrelTransform: {fileID: 7442443419788325056}
@@ -485,6 +489,37 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 10.346319, y: 2.6368096, z: 5.5847664}
   m_Center: {x: 0.006264448, y: 1.2667948, z: -1.8277111}
+--- !u!114 &-1355217651115294554
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5848915433920635072}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5fea29bb7c508c244a1f805a5fd3fc4d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  outlineMode: 0
+  outlineColor: {r: 1, g: 1, b: 1, a: 1}
+  outlineWidth: 2
+  precomputeOutline: 0
+  bakeKeys: []
+  bakeValues: []
+--- !u!114 &4281797126201779825
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5848915433920635072}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d3e01160669f82e499d3f5c9fdb4eebc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  selectionColor: {r: 0, g: 0.30980393, b: 1, a: 1}
 --- !u!1 &7765907135307089973
 GameObject:
   m_ObjectHideFlags: 0
