+++ Red Strike/Assets/VehicleSystem/Vehicles/Trike/Trike.prefab
@@ -45,6 +45,8 @@ GameObject:
   - component: {fileID: 1493612783217301255}
   - component: {fileID: 19923418697428257}
   - component: {fileID: 2436856578806251895}
+  - component: {fileID: -1316650565147742908}
+  - component: {fileID: -7762954015810627755}
   m_Layer: 6
   m_Name: Trike
   m_TagString: Vehicle
@@ -184,6 +186,37 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1.4111015, y: 1.1139936, z: 3.0195768}
   m_Center: {x: 0, y: 0.70709383, z: 0.037513375}
+--- !u!114 &-1316650565147742908
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2231230407424723898}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5fea29bb7c508c244a1f805a5fd3fc4d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  outlineMode: 0
+  outlineColor: {r: 1, g: 1, b: 1, a: 1}
+  outlineWidth: 2
+  precomputeOutline: 0
+  bakeKeys: []
+  bakeValues: []
+--- !u!114 &-7762954015810627755
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2231230407424723898}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d3e01160669f82e499d3f5c9fdb4eebc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  selectionColor: {r: 0, g: 0.30980393, b: 1, a: 1}
 --- !u!1 &2831294200106519584
 GameObject:
   m_ObjectHideFlags: 0
